price-service/src/test/java/com/account/collateral/service/PriceServiceTest.java
package com.account.collateral.service;

import com.account.collateral.model.AssetPrice;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.MockitoJUnitRunner;

import static org.junit.Assert.assertEquals;

@RunWith(MockitoJUnitRunner.Silent.class)
public class PriceServiceTest {

    @InjectMocks
    private PriceService priceService;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    public void getAssetPriceTest(){
        AssetPrice[] assetPrices = priceService.getAssetPrice();
        assertEquals(assetPrices.length, 4);
    }
}
